Movement preliminaries
Added preliminaries for moving to the game. Players have age, shown in
purple. Can switch between "move" and "interact" commands. Added help
for gameplay. Fixed some stuff with loading and saving. Code was very
broken with regards to column and row (was switched up all over the
place and miraculously working) so that's fixed too. Holiday grid
is now vertical as a result. In motion selection, added ghost animation.
Advancing timestep should move player, but this is not in the game yet.